绑定系统用户权限，绑定系统用户角色，系统菜单编辑界面功能完善

+++ SPring_Cloud_Group/spring_cloud_uc_server/src/main/java/com/rjpa/repository/LzhAdminPermissionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface LzhAdminPermissionRepository extends JpaRepository<LzhAdminPermissionEntity, Integer> {
-    LzhAdminPermissionEntity findById(long id);
+    LzhAdminPermissionEntity findById(int id);
 
     List<LzhAdminPermissionEntity> findByPermission(String permission);
